[IMP] calendar_event_without_group: If a user has not created a meeting, he can not accept invitations, or cancel.
