+++ calendar_event_without_group/models/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 # (c) 2016 Alfredo de la Fuente - AvanzOSC
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
-from . import calendar_event
+from . import calendar
